app/requirements.txt
gunicorn==20.0.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
setuptools>=40.3.0
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mkl==2019.0
numpy==1.18.5
olefile==0.46
Pillow==4.3.0
protobuf==3.12.2
psycopg2>=2.6.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
torch==1.5.1
urllib3==1.25.9
Werkzeug==1.0.1
opencv-python-headless==4.2.0.34
google-cloud-logging==1.15.0
